requirements.txt
asn1crypto==1.5.1
astroid==3.3.4
aws-lambda-powertools==3.0.0
boto3==1.35.37
botocore==1.35.37
click==8.1.7
dill==0.3.9
isort==5.13.2
jmespath==1.0.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
pg8000==1.31.2
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
Pygments==2.18.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.9.2
ruff==0.6.9
s3transfer==0.10.3
scramp==1.4.5
shellingham==1.5.4
six==1.16.0
tomlkit==0.13.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
